fix bugs

+++ QuanLyCuaHangBanSach/DTO/CustomerBillDetailDTO.cs
@@ -14,7 +14,7 @@ namespace QuanLyCuaHangBanSach.DTO
         }
 
         public CustomerBillDetailDTO(
-            int maDonHang, int maSach, int soLuong, double donGia
+            int maDonHang, int maSach, int soLuong, decimal donGia
         ) : base(maDonHang, maSach, soLuong, donGia)
         {
         }
@@ -24,7 +24,7 @@ namespace QuanLyCuaHangBanSach.DTO
             this.MaDon = (int)row["maDonKhachHang"];
             this.MaSach = (int)row["maSach"];
             this.SoLuong = (int)row["soLuong"];
-            this.DonGia = Convert.ToDouble(row["donGia"]);
+            this.DonGia = Convert.ToDecimal(row["donGia"]);
             this.ThanhTien = this.SoLuong * this.DonGia;
         }
     }
